Revert back to standard dependency version requirements (#9)
Using the caret ^ for versioning means that the plugins requires poetry >=2.0.0,<3.0.0. This means that one cannot simply just upgrade the plugin then update poetry when coming from poetry 1.8.3. The plugin needs to be removed and reinstalled once poetry is updated. This version makes it easier. Upgrade the plugin poetry `self add poetry-pre-...@latest`

It also does not break the compatibility continuity of the plugin.